Remove zmq_{send,recv}msg, make zmq_{send,recv} do their jobs.
So, in the version of ZMQ installed on my Ubuntu (ZMQ 2.20), there are no
functions sendmsg/recvmsg, only send and receive, but with signatures for
sending zmq_msg_t objects. So to get things working for me, I've renamed this.
If there are versions with and without these functions in widespread use, we
may have to do something terribly clever to get things to work properly...

There's one broken test in the basic test file (checking for the RCVMORE flag
in getopt), but it's not entirely clear if that's due to my changes or
something else, since I can't run the previous version of the code with my
libzmq.